Reorganized and updated.
Moved the pricefeed file into the correct place.
updated the docker-compose file to use the new name.
Added a .vscode directory with a launch.json file to make debugging easier.
Added a .vscode/settings.json file to make the python linter happy.
Added a .gitignore file to ignore the .vscode directory.
Updated the Dockerfile to use the new name.
Updated the poetry.lock file to use the new name.
Updated the pyproject.toml file to use the new name.
Updated the __init__.py file to use the new name.

Signed-off-by: Brian of London (home imac) <brian@v4v.app>